修改Receipt中departmentId类型为Integer

+++ common/src/main/java/exort/api/http/review/entity/receipt/AssociationMemberReceipt.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AssociationMemberReceipt {
 
-    Long departmentId;
+    Integer departmentId;
     String message;
 
 }
